Debugger: Save/restore IP correctly in SingleStep mode
Summary:
SingleStep mode should save and restore the IP correctly just like the
full interpreter loop. This wasn't a visible problem before because the
only time we single-step is when throwing, but as we add more
instructions that can use single stepping, it should be possible for
them to save/clear the IP knowing that it'll be restored on exiting from
SingleStep mode.

Note that this change makes no difference to multi-stepping mode,
it just makes SingleStep mode behave properly.

Reviewed By: avp

Differential Revision: D73798413

fbshipit-source-id: 0de3841d99beb5bd125a9ef9f8ef352a3d95edae