Fix minor non-determnism in shermes C-program generation.
Summary:
The next diff in this stack adapts the bytecode determinism test to
ensure that the -emit-c option of shermes produces a deterministic C
program.  I discovered that it didn't -- there were minor order
differences in a switch statement introduces as part of exception
handling.  This diff makes that order deterministic.

Reviewed By: avp

Differential Revision: D74351758

fbshipit-source-id: 721e361a1dd77e1d47f218876b199f923c3e75a7