VM: Extract function to check if a double can be inlined into SHV
Summary:
Make the "inlinability" condition explicit in `SmallHermesValue` by
exposing a function that can be queried to centralize this logic.
This is useful outside `SmallHermesValue` as well so that the JIT can
check whether doubles can be made into `SmallHermesValue` without any
allocations.

`canInlineDouble` is a public API, so implement it on the Adapter as
well.

Reviewed By: neildhar

Differential Revision: D71980383

fbshipit-source-id: 3141e03d497454ac335f4589e75a8f011b7dfc62