JIT: Simplest inline value population in NewObjectWithBuffer
Summary:
This is the first diff in an incremental path towards avoiding having to
reparse the value buffer when we run NewObjectWithBuffer from JITted
code. It supports HV32 and HV64.

For now, we also have a `newObjectWithBufferSlow` variant to which we
bail if during compilation we detect that we're not able to use this new
path. Eventually we'll support everything in `newObjectWithBuffer` and
we can delete `newObjectWithBufferSlow`.

Not yet supported:
* Boxed doubles in HV32 mode (inline doubles are supported)
* Strings (requires exposing the IdentifierTable entries to the JIT)

Reviewed By: tmikov

Differential Revision: D71652357

fbshipit-source-id: 26c9d18645cf1eeb546dcc3bb6e7ec653ec523a0