VM: Allocate JSObject and property storage together
Summary:
For object literal purposes, it's nice to be able to ignore the write
barrier after creating a new object while we populate the values.
This can only be done if the memory to which we're writing is in the
young gen, so this diff allows us to make guarantees about that.

Add a GC API to allocate memory for two objects at once.
This lets us allocate both the JSObject and its property storage
together if the total size is able to fit into the young gen.
Use that API whenever possible.

MallocGC also needs to be handled, so just call malloc twice for that one
because there are no generations and mark both allocations as
`inYoungGen`. `canAllocateInYoungGen` will always return `true`.

Reviewed By: tmikov

Differential Revision: D72205121

fbshipit-source-id: 0c297f47118a9f8fcae6667008693bda3fd80885