Add TransparentConservativeVector for IdentifierTable
Summary:
In order to allow the JIT to read fields in the IdentifierTable
for faster string value population during object literal creation,
transition from a custom subclass of `std::vector` to a new
`TransparentConservativeVector` which has similar memory
characteristics. I tried to implement it from scratch, but it ended up
being a slight regression on certain benchmarks like earley-boyer,
so instead this diff extends `llvh::SmallVector` which has a known
offset for the data pointer at `0` and the member is protected so it's
easy to expose the offset to the JIT.

Like the existing `ConservativeVector`, it wastes less space in the
common case where the number of elements created dynamically is small
compared to the number of elements known at a given reservation time.

The pointer is stored at offset 0, so it is accessible directly with
`&`.

Reviewed By: tmikov

Differential Revision: D72091373

fbshipit-source-id: 4fe7ef3ebdade7b3a6f6f3fb317459214e689380