JIT: Handle strings in newObjectWithBuffer
Summary:
We can force the strings to be allocated at JIT time for the whole
function, so we can simply read them directly from the IdentifierTable's
lookup vector because we know the SymbolIDs at JIT time too.

So now strings don't need to fail the `newObjectWithBuffer`
implementation back to the slow version.

The JIT compilation also needs to capture the IP now because it can
allocate. Only do this when necessary by splitting up the JIT
compilation API and directly checking for the JITCompiledFunctionPtr in
the Interpreter. With non-allocating `shouldCompile` function, we can
check if compilation is needed without actually saving the IP.

Reviewed By: tmikov

Differential Revision: D72091374

fbshipit-source-id: eddb1747024fb47b3a27c0fcf5ffc29d45b9ce01