Fix `hermes-parser-wasm` CMake target (#1702)
Summary:
- Fixes CMake target for `hermes-parser-warm`
- This change also re-writes some of `build.sh` to be more cross-platform friendly, as the `rename` utility is not part of MacOS standard lib, and the `rename` binary that can be installed has a different API.
- The `--parents` option is also unsupported on MacOS's `cp`, and rather than going down the rabbit-hole of GNU `coreutils` compatibility, I used what I believe to be more cross-platform friendly bash.

While working on `flow-api-translator`  I was unable to build `hermes-parser-wasm` as the CMake target no longer built without error.

Pull Request resolved: https://github.com/facebook/hermes/pull/1702

Test Plan: It appears that `test-emscripten` job in CircleCI was removed at some point, and there is no job to build this in GitHub workflows. Happy to add a job back in but before I go down that path, it's possible that there was good reason for removing it, so rather than just putting it back, my testing strategy is "I built it successfully" (Which does not exactly instill the most confidence in ensuring it remains working)

Reviewed By: avp

Differential Revision: D74429774

Pulled By: pieterv

fbshipit-source-id: d0cece4dda3e7df29e52928a68a7a52f0b5e8c3e