Delete mmap hint when allocating storage
Summary:
We currently provide a random hint when allocating storage to randomise
where our heap segments are allocated. However, this can cause problems
if the hint results in a mapping that conflicts with the reported stack
bounds. On linux, the main thread stack does not have the full virtual
address space allocated upfront. So if a mapping is created in the
region associated with it, it can invalidate our cached stack bounds.

We could try to restrict the hint to regions that we know will not
conflict with the main thread stack, but for now, it is simpler to just
delete the mechanism, since we no longer need it.

Reviewed By: lavenzg

Differential Revision: D74619999

fbshipit-source-id: 2cacf7123fff2e75065ccaf3305de6ea4ca6e0e4