Check for recursion depth exceeded in SemanticResolver
Summary:
`visitESTreeNode` or `visitESTreeChildren` can fail due to exceeding the
recursion depth. If that happens, the semantic information on the
children of the current node won't be valid. In the cases where we then
rely on the information being valid, we must first check that
`recursionDepth_` is not `0`. If it is `0`, we can't do anything more
and we should just return as quickly as possible to make sure we
report the error and don't crash due to invalid semantic info.

I've added this check in all the places I could find in this diff that
may rely on it. There may be one or two that are not strictly
necessary, but it's easier to be defensive about this.

Reviewed By: tmikov

Differential Revision: D74010361

fbshipit-source-id: fbbeaf597a43f3099678aaf5a9e64e95030933c7