Add intermediate state for the Hermes Root API
Summary:
This diff adds the Hermes Root API interface, but keeps the static
functions within HermesRuntime still. This is a necessary intermediate
step for RN ([see failing tests here](https://github.com/facebook/react-native/actions/runs/14800547982/job/41558052598?pr=51082)), which depends on the version of
Hermes that is on `main`, which does not have the Root API yet.

Once this lands, we can land the next diff in the stack that migrates
all the usages to `makeHermesRootAPI`, and remove all the static methods
from `HermesRuntime`.

Reviewed By: neildhar

Differential Revision: D74594455

fbshipit-source-id: 7db62051082e3384b0f5215a2262cda58208db9f