VM: Add epoch to Runtime and cached parent flag to JSObject
Summary:
In order to properly invalidate `AddPropertyCacheEntry`, we need to track
when objects used as parents are modified.
To handle this, we store two pieces of data.

`isAddPropertyCacheParent` in `JSObject` indicates that the object is in
a prototype chain of an object which has been cached in
`AddPropertyCacheEntry`.

`parentCacheEpoch` in `Runtime` will be incremented every time we modify
any object `p` with `cachedParent` set in a way that invalidates the add
property caching for any object `o` which has `p` anywhere in its
prototype chain. When it hits its maximum value, we just invalidate all
the add property caches.

Epoch incrementing operations for isAddPropertyCacheParent objects:
* Changing the HiddenClass
* Changing the parent

Reviewed By: tmikov

Differential Revision: D73394015

fbshipit-source-id: 510f9fc55ee5f364863528660801773d0824858b