VM: Store AddPropertyCacheEntry in RuntimeModule/SHUnit
Summary:
We want to have a data structure in `RuntimeModule` to back the index
stored in `WritePropertyCacheEntry`, so store `AddPropertyCacheEntry`
there.

This new entry stores the starting class for the transition, the result
class, and the object's parent and parentEpoch (D73394015) at the time
of caching. If these change, the cache is invalidated.
The slot in which to place the new property is stored in the bottom byte
of the parentEpoch, which isn't going to be too big in general.

This way, the `WritePropertyCacheEntry` can be used to cache both
modification and addition at the same time.
We also reserve the first element and never write to it, ensuring
it's always a cache miss and avoiding an extra check in the cache hit
path.

`SHUnit` has similar structured, placed in `SHUnitExt`.

The `AddPropertyCacheEntry` is never freed once allocated, because it
doesn't conflict with caching modifications that don't add properties.

Reviewed By: tmikov

Differential Revision: D73520781

fbshipit-source-id: 796ddae5e105427c5f1a34efec777ae9c7348f62