VM: Allow WritePropertyCache to store an addCache index
Summary:
We want to be able to cache add property transitions between
HiddenClasses in this structure, but we don't want to increase the size.

Reduce the space taken by the cache index to 8 bits.

The remaining 24 bits are used to encode an index, which will index into
a list of `AddPropertyCacheEntry` in the `RuntimeModule` (D73520781).
The `0` entry in that list is reserved, so a `0` addCache index here
will never be written to (it'll always be a cache miss).
This way we don't have to check for validity on the cache hit path.

`SHWritePropertyCacheEntry` has the field name modified, but nothing
was actually using the field directly so no modifications required
there.

Reviewed By: tmikov

Differential Revision: D73520780

fbshipit-source-id: c611ed8c279e5c696a1e7f716a0c52e834bcaacd