VM: Centralize HiddenClass access in JSObject
Summary:
We want to be able to perform checks on whether the parent cache epoch
should be incremented for PutById add property caching.
We're going to increment when HiddenClass changes, so we need to be able
to have all sets of the HiddenClass go through a single function.

Rename the field and document it to make this clear.

Go through all accesses to the field and change them to use the helper
methods. This has the slight advantage of also removing some unnecessary
null checks, as the `clazz_` was always required to be non-null but
there were certain set/gets that were not using the `NonNull` variants.

Reviewed By: tmikov

Differential Revision: D74667465

fbshipit-source-id: e7b3b9fc061ed5dbe1471186334be9d3eed4aebf