VM: Use PutById add property cache in interpreter/JIT
Summary:
Actually read the `AddPropertyCache` entry in the interpreter and the
JIT PutById paths.

In order to hit the transition cache, we need:
* A cached result class
* An add cache entry that's "in use"
* The start class to match the object's current class
* The parent epoch to currently match the cached epoch
* The parent of the object to match the cached parent

If those checks pass, we actually add the property.
I tried to place the two non-allocating paths inline in the interpreter,
but that didn't actually improve performance on any benchmarks, and it
did cause a notable (1-2%) regression on certain benchmarks like crypto
that almost never actually hit this transition cache due to having too
much code inlined in a hot path in the interpreter loop.

It would be nice to avoid having to decode and reencode the `resulClazz`
compressed pointer, but if we're storing it into a strong reference we
must invoke the read barrier on the `WeakRoot`.

Reviewed By: tmikov

Differential Revision: D73520785

fbshipit-source-id: 98fb12d1bcd7db348abb9a639d3b9c2506ce3a22