Make `OrderedHashMap::has/get` take in a `HermesValue` over a `Handle`
Summary:
While working on the Set methods, we realized that `OrderedHashMap::has`
and `OrderedHashMap::get` takes in Handle for the key. However, nothing
is done in these methods that warrant a `Handle`.

Diff removes the `Handle` usages and simply takes in a `HermesValue`, and
changes current call sites. Later, the changes to add more Set methods
will also use these APIs, and we don't have have to store theses
values in `Locals`

Reviewed By: fbmal7

Differential Revision: D74844756

fbshipit-source-id: 3289ca51ba852f69596f8782a28fcf9a3742bc51