Remove GCScope in array push and pop fast paths
Summary:
Delay creating the GCScope until after the fast path, since it
contributes significantly to the cost of these functions.

It's not clear that we actually need a GCScope here or if a marker would
suffice. But given that we are in the process of deleting GCScopes, and
that the cost can be contained to the slow path, it doesn't seem worth
trying to eliminate right now.

Reviewed By: avp

Differential Revision: D75037012

fbshipit-source-id: e1ee64e786b786bd806b400c9a53d466544a9003