Remove scratch register
Summary:
Instead of allocating a scratch register in every frame purely for the
use of BoundFunction calls, have _boundCall allocate a local.

This allows us to delete the scratch register. Combined with upcoming
support to eliminate the DebugEnvironment register, this will mean we
won't need any registers at the start of the callee frame.

Reviewed By: tmikov

Differential Revision: D74947040

fbshipit-source-id: 92fc1aa22dee76034e7273b3035c35d70c296c57