Use inline assembly to initialize stack on armv8
Summary:
We currently force an out-of-line call for initializing registers in
order to prevent the compiler from emitting an indirect call into libc.
This is a fairly hot path, so making the call is not ideal. Address this
by adding an inline assembly implementation for armv8. We can gradually
add other inline assembly implementations to eliminate the call
altogether.

Reviewed By: tmikov

Differential Revision: D74503660

fbshipit-source-id: ff1a34b2d78c5aded5d1f8c29db1965f9ed5146b