Use left paren location for start of object literal methods
Summary:
The convention for these locations is to place them at the `(` because
that way it's easier for tools consuming this information to, e.g.,
reformat text. The surrounding PropertyNode already has the actual start
location of the method name.

The one complication is that using the `(` as the start of the
FunctionExpressionNode makes lazy compilation fail due to attempting to
parse a property node from the start of the FunctionExpressionNode.
We already have a mechanism to deal with this for classes, so update the
`getFunctionRange` function accordingly and ensure the parent is passed
along for object literal keys as well.

Reviewed By: tmikov

Differential Revision: D75537251

fbshipit-source-id: e95ae892803bfef86ae10bcb2a8f2035ff62c543