Always populate SavedIP in caller
Summary:
We currently sometimes skip storing the current IP before making a call,
and instead rely on the callee to populate it with
`Runtime::saveCallerIPInStackFrame`. This can be inefficient, since the
callee ends up unconditionally having to saving the IP, even if the
caller had already saved it.

Instead, have the caller always save the IP, which allows us to
eliminate saving in the callee in the next diff.

Reviewed By: avp

Differential Revision: D75115009

fbshipit-source-id: c3b009d01c9cd50dd76e815b0b9e3a60393ad5f5