Add implicit flag to BlockStatementNode
Summary:
In order to make the Parser more compliant from the perspective of
non-compiler users, we want to be able to strip artificial
`BlockStatementNode`s that didn't exist in the original source.
We still need to insert them because the `SemanticResolver` would like
to rely on `IfStatement` always having the block as a
consequent/alternate, but other clients don't need to keep it around.
Adding a flag makes it easy for consumers of the AST to know if the node
was in the input source.

Reviewed By: tmikov

Differential Revision: D75610190

fbshipit-source-id: 332767e0f3cfe2176690b8c283917b7ad14b2c03