Add jitCall for each CellKind
Summary:
Add a new array similar to VTable that stores the "jitCall" function for
each object type. This allows the JIT to quickly find the right handler
for calling a given object.

Calling through this "vtable" function is preferable to going through
`_jit_dispatch_call` as we do right now, because instead of making an
indirect call to `_jit_dispatch_call`, we can make an indirect call to
the actual implementation for the callable. This means less work needs
to be done on each call, and branch prediction for each callsite is
specialised to the type of function it actually calls.

This diff just introduces jitCall and adds a default implementation that
throws. Subsequent diffs will actually implement and use this.

Reviewed By: avp

Differential Revision: D74750481

fbshipit-source-id: 78cd9aae5b87bd24cac6786a97966643ae3e4da5