Call through jitCall functions in the JIT
Summary:
Instead of calling `_jit_dispatch_call`, directly invoke the logic for
handling the specific `Callable` through the jitCall array.

Calling through the "vtable" function is preferable to going through
`_jit_dispatch_call` as we do right now, because instead of making an
indirect call to `_jit_dispatch_call`, we can make an indirect call to
the actual implementation for the callable. This means less work needs
to be done on each call, and branch prediction for each callsite is
specialised to the type of function it actually calls.

Reviewed By: avp

Differential Revision: D74750479

fbshipit-source-id: f5cc4415300ee431a35d3d00cafd8d33eb09928f