Speed up array fast path prototype check
Summary:
Tighten the requirements on the prototype chain of an array by requiring
that the immediate parent is the original one, and that no other objects
in the prototype chain have indexed storage. This allows us to eliminate
the virtual call to `getOwnIndexedRange` for each parent.

This also allows us to use the epoch to skip checking anything after the
immediate parent, which we can consider separately.

Reviewed By: avp

Differential Revision: D75562680

fbshipit-source-id: 47c6d7ec0a14d517574d991bc3a17cb61256d8de