Add a "good change" counter to the read prop cache
Summary:
The "good change" counter counts how many times a
`ReadPropertyCacheEntry` has been updated with new cache information.

The intent is to use it to guide JIT fast path decisions.

Note that the counter does not currently count cases where the cache
missed or when something couldn't be cached. The reasoning is two-fold:
- We currently don't invalidate the cache on misses or failures, so it
would be difficult (but not impossible) to count these transitions.
- Even if the cache doesn't hit every time, it is probably a good idea
to inline the fast path if it has changed only once.

Renamed the name of the slot field to avoid errors, since it had to be
shrunk to 16 bits.

Reviewed By: avp

Differential Revision: D75386614

fbshipit-source-id: 6adeca37f68ffabb0374a61ac3c6e6951bd50204