JIT: add lazy HC id-s.
Summary:
Implement lazy assignment of unique 16-bit identifiers to hidden
classes used by the JIT. This enables the JIT to emit efficient code
for checking hidden classes by loading the ID from the class and
comparing it with a 16-bit immediate, rather than comparing
GC-managed pointers which would require more complex code generation
and management since hidden classes are moveable.

The implementation adds a `lazyJITId_` field to HiddenClass and
assigns IDs only when the JIT first encounters a hidden class during
compilation.

The JIT tracks all hidden classes it references in an
instance of `ArrayStorageSmall` to ensure they are properly marked. This
adds an interesting wrinkle in that performing a GC allocation might
move the HC while we are working with it in the GC. Fortunately in
practice we don't keep pointers to HCs around.

Used in the next commit.

Reviewed By: avp

Differential Revision: D75195054

fbshipit-source-id: a242bceca4e9e5fe2940fe8e30332d30e43fb8f0