JIT: Inline allocation in newObjectWithBuffer
Summary:
Now that we can inline allocations into the JIT, we can allocate and
populate the object without any calls.
Add a new `alloc2InYoung` which allows for allocating 2 cells in the
young gen (JSObject and PropStorage) and use that if there's indirect
property storage needed.

Allocation can result in a slow path so we have to add one.

Reviewed By: neildhar

Differential Revision: D74770706

fbshipit-source-id: fc02bf689d50a3ae6cbdf44446f2203cb7a562ce