Disable blocking collection during YG collection
Summary:
Blocking OG collections during YG collections are currently broken as
they can cause the compactee to transition into the "evacActive" state,
which means that we might start evacuating it halfway into the YG
collection. However, only parts of the compactee are evacuated, and
the surrounding YG collection code cannot deal with this.

As an immediate fix, disable blocking OG collections until we rearrange
the surrounding code.

Reviewed By: dannysu

Differential Revision: D76322544

fbshipit-source-id: 4d6e183fd4b097e62741c78189b5fcc6fa7f2fec