Remove transition check in HiddenClass
Summary:
We have an optimisation to avoid materialising a DictPropertyMap for an
HC if the property being looked up is in the transition map, since that
means that the property definitely does not exist. However, this
complicates the API since we also need to pass in the expected flags for
the transition.

Given the very small number of hidden classes that are actually created
in a typical workload, the savings from not created the DictPropertyMap
seem small. Remove this mechanism to simplify the API.

Reviewed By: avp

Differential Revision: D74077355

fbshipit-source-id: b737474145f4200d21c0eff6ac483270f5bb2ccd