Remove newTarget from NativeArgs
Summary:
Instead of explicitly storing a pointer to the newTarget in NativeArgs,
compute it relative to the arg pointer we already have.

This is much less important now that we create the `NativeArgs` locally
in each `NativeFunction`, but it still helps the remaining cases where
we pass `NativeArgs` as an argument to a function. For instance, we have
helpers in `Array.cpp` that do this.

Reviewed By: avp

Differential Revision: D75895960

fbshipit-source-id: 6bef5b086cfc6e77e8700b89073034f14f163411