GC: Avoid duplicating memory instrumentation logic in MallocGC
Summary:
The MallocGC implementation of `make2YoungGenUnsafe` just calls through
to `makeA`, which allocates the object IDs and registers allocations
with the instrumentation.

The Hades version does not allocate the object IDs, leaving that to the
GCBase code.

This means that the MallocGC version runs the code twice, but the Hades
version runs it once. Fix this by making MallocGC use `makeAImpl`
instead, allowing us to leave the allocation ID logic in GCBase.

An alternative would be to move the allocation ID calls into each
individual GC, but this is easier (and only changes MallocGC).

Reviewed By: neildhar

Differential Revision: D76445712

fbshipit-source-id: 390195ddc77ab231894dc56238ed1b3d50f2bc9c