Skip restoring IP when unwinding C++ exceptions
Summary:
If Hermes is built with exception on OOM, we may unwind through the
interpreter when an exception is thrown. If this happens, the register
stack will not have been popped, so it is invalid to try and obtain the
saved IP.

Instead, we should just skip restoring the IP in the runtime, since the
runtime is about to be torn down anyway.

Reviewed By: avp

Differential Revision: D76473687

fbshipit-source-id: 3a7e591d29b0c30e3d958778c7fd648c45e06054