Defer registering SamplingProfiler until all fields are set
Summary:
We currently register the `SamplingProfiler` instance with the profiling
thread in the base class constructor, which means that the fields in the
platform specific subclass may not yet be initialised. This can cause a
race with the profiling thread on the uninitialised fields, causing a
crash.

Move the registration after the fields are initialised, so the profiling
thread does not attempt to read any fields until after they are all
initialised.

Differential Revision: D76539324

fbshipit-source-id: 30417e08d9513de6514040c2bbd47b1f68ab0af0