Avoid libc call when pushing onto ArrayStorage
Summary:
Calling `GCHVType::uninitialized_fill` causes the compiler to emit a
libc call to initialize the empty slots. However, this is pretty
inefficient since we know that in practice the number of additional
elements is small and does not justify the indirect call.

Instead, manually force the compiler to emit a loop with individual
stores.

Reviewed By: avp

Differential Revision: D76298410

fbshipit-source-id: 85c559a9633721ad33d429bf51689fae67e1a771