Don't let compileJS or CompilerDriver set full debug info
Summary:
`compileJS` and `CompilerDriver` are not meant to be entry points to
debugging. Therefore, they should not be allowed to be set full debug
info. However, it may still be useful to be able to see the IR / bytecode
which is generated for full debug info. So we create a new `Xg3` flag.

Reviewed By: avp

Differential Revision: D76168304

fbshipit-source-id: 70489a0b5c83db29da4cb9d4fb9405a4bf842c46