Disable trimming ArrayStorage
Summary:
Trimming ArrayStorage can be confusing, since the capacity of a storage
can change when an unrelated allocation occurs. It is also rarely used
in practice, since Hades in concurrent mode will only do trimming when
sweeping occurs on the mutator.

This diff just turns off trimming, while retaining the capability in the
GC, as well as the supporting code and tests for ArrayStorage.

Reviewed By: avp

Differential Revision: D75650459

fbshipit-source-id: 90a87cd0b6aaf88c91aa57f0dcc49c0053cdad7f