VM: Inline functions in Operations.h
Summary:
There's a few of these small functions that are trivially made inline.

They are generally just wrappers for other APIs that are named according
to the spec, or just perform some cheap operation before calling a
function that does real work.

It should be possible to do this for even more functions, but there's
cyclic header issues at that point so this is just the easiest functions.

This makes the `hermes` CLI binary a few bytes smaller, and has no
visible impact on performance that I can find.

Reviewed By: fbmal7

Differential Revision: D76827114

fbshipit-source-id: 06d175dfa95abc43e5eb00524e89334e53dbc951