Reduce usage of CompacteeState::evacActive
Summary:
`evacActive` is currently used in several places as a shorthand for
"there is a compactee and marking is completed". However, this is pretty
confusing, and it would be clearer to explicitly state the collection
phase instead.

This is also useful in later diffs, which eliminate the evacuation
tracking from `CompacteeState` altogether.

Reviewed By: lavenzg

Differential Revision: D76322545

fbshipit-source-id: 38104800607226b0d6bdacd260c8e7414a03d383