Fix assigning private property to lazy objects
Summary:
The code path for adding private properties does not currently check
whether the object is lazy. This means that it ends up adding properties
to the lazy hidden class, instead of first materialising the lazy
properties and then adding the private property.

Fix this by checking for lazy objects in `defineNewOwnProperty`.

Reviewed By: avp

Differential Revision: D76765266

fbshipit-source-id: 2485681e8fad7bb35c8bf57f9b95e361ec19df4d