Use DoEmitDeclarations::Yes for synthetic functions
Summary:
We should always be passing in `DoEmitDeclarations::Yes` unless it's
incorrect to do so. In practice, currently these synthetic functions
never have declarations in them so this is a no-op. But in future diffs,
emitScopeDeclarations will also be responsible for associating
`LexicalScope`s with `VariableScope`s, so we want that association to
happen for the synthetic functions as well.

Reviewed By: avp

Differential Revision: D76077259

fbshipit-source-id: 909718689af39d51751138a6d8280617e76ea12a