Fix ASAN build issues on Linux with Clang
Summary:
On Linux, when we build with ASAN/UBSAN, libhermesvm.so does not have
dependency on the ASAN/UBSAN library. This is different from MacOS,
where Apple Clang adds the dependency. As a result, when shermes is
compiling the executable, it fails to link due to undefined ASAN/UBSAN
symbols. This diff adds the sanitizer flags to SHERMES_CC_SYSCFLAGS and
SHERMES_CC_SYSLDFLAGS when the CMake build is configured with ASAN/UBSAN.

Reviewed By: neildhar

Differential Revision: D76912416

fbshipit-source-id: 96d447d59620009d1e6c9c5d177301411369c49e