Simplify initialisation of register stack
Summary:
Remove the inline assembly implementation for initialising the register
stack, in favour of using an empty asm statement that forces the
compiler to keep the loop intact.

While here, also remove the constant checks, because we now almost never
call this function with a constant value, and when we do, it is <2, so
the compiler should do a reasonable job optimising it.

This does generate slightly worse code on arm64 than the inline
assembly (at least in the interpreter), but it solves the primary
problem of the out-of-line call, so it's not clear that it is worth
keeping the assembly.

Specifically, on armv8, there is now an extra mov and add in the loop:

Loop body before:
```
stp     q0, q0, [x0], #32
cmp     x0, x1
b.lt    .Ltmp1
```

Loop body after:
```
add     x8, x0, #32
stp     q0, q0, [x0]
cmp     x8, x1
mov     x0, x8
b.lo    .LBB1_2
```

https://godbolt.org/z/Pvhh59qs5

Reviewed By: avp

Differential Revision: D76523008

fbshipit-source-id: d70a7a3f0f73d90759918ba6c0c5d9b99b1b6e59