Fix missing read barrier in JIT property cache population
Summary:
The barrier cannot be skipped if we are reading the value in order to
store it in a strong reference somewhere. In this case, since we are
reading the `WeakRoot` in order to store it in the `usedHCs` storage, we
must use a barrier to inform the GC that the HC is live. Otherwise, the
GC may collect the class while `usedHCs` still holds a reference to it.

Reviewed By: tmikov

Differential Revision: D77191681

fbshipit-source-id: 515d330ef9633d96de4b5797ce8cadfb04fb83d6