Remove `MutableHandle &` in JSObject
Summary:
## MutableHandle &
Passing a `MutableHandle &` is redundant, since it already is a pointer,
and since we recently added a copy constructor. So, change all cases to
pass by value. Additionally, use the `.set()` methods of `MutableHandle`
to avoid unused parameter warning.

## tmpSymbolStorage

One side effect of this is to clean up a long-standing mess. In some
rare cases - namely when obtaining a descriptor that we need to pass to
HostObject or proxy - we need to allocate and store a SymbolID in the
descriptor, and we need a way to keep that symbol alive for the lifetime
of the descriptor.

Previously we achieved that by passing
```
        MutableHandle<SymbolID> & tmpSymbolStorage,
        ComputedPropertyDescriptor &desc
```
to the function that potentially allocating the symbol. Additionally we
passed the symbol storage to all functions consuming the descriptor, in
order to force the two lifetimes to be the same. Needless to say, this
was clumsy. Worse, it forced us to allocate a `MutableHandle<SymbolID>`
even in cases where it wasn't needed.

Clean this up by introducing:
```
class ComputedPropertyDescWithSymStorage : public ComputedPropertyDescriptor {...};
````

This encapsulates both the descriptor and the MutableHandle, ensuring
their lifetimes are the same. Additionally, this forces us to clarify
which functions actually need this.

This resolved in unexpected simplifications.

## Misc

Lastly, while we were visiting various files, we made minor cleanups
like deleting unnecessary handles that simply used to alias
PinnedValue-s.

Reviewed By: avp

Differential Revision: D77080893

fbshipit-source-id: 0f4e8183f9be056baeb1bb0e20c072fc2aca3b19