HermesValueTraits: fix shocking bug with StringPrimitive
Summary:
ExternalStringPrimitive was not recognized as a string and consequently
was encoded as an object instead of string. This wasn't an issue because
vmcast<> uses `getPointer()` and checks the CellKind, ignoring the
specific HV tag.

But it can be an issue if someone used `getString()` on a value encoded
this way.

What is even more confusing is that ExternalStringPrimitive::create()
returns a HermesValue with the correct String tag. Decoding and
re-encoding it changed the tag to `Object`. So, if we used
`.castAndSetHermesValue()` to eliminate the re-encoding, we ended up
with a String instead of Object tag. Subtle indeed!

One might argue that the traits system is too complex and that we don't
need so many HV tags, but for now just fix this.

Reviewed By: avp

Differential Revision: D77094256

fbshipit-source-id: 0ceb34197307320d26249bc0602dd6f5c62678de