Use ID_Start property instead of letter property for Unicode IDs
Summary:
JS identifiers are supposed to start with `ID_Start` Unicode codepoints,
not only letters.

Unicode spec defines `ID_Start` as:
```
#      Lu + Ll + Lt + Lm + Lo + Nl
#    + Other_ID_Start
#    - Pattern_Syntax
#    - Pattern_White_Space
```

So I added the tables for `Other_ID_Start` and a combination of
`Pattern_White_Space` and `Pattern_Syntax`, along with an extra check.

This should never happen in practice so the minor performance
hit shouldn't matter, and these extra tables only add 5 extra ranges to
the file.

The next steps include doing the same thing for `ID_Continue`.

Alternate approach I didn't do:
We could just one new table to
`UnicodeData.inc` that only contains the `ID_Start` points. There are
only 740 ranges so this takes less than 6 KB, but in reality the
`ID_Start` table is almost identical to the letters table so the above
approach is quite simple.

Reviewed By: neildhar

Differential Revision: D77245698

fbshipit-source-id: d68e867e16d031f9608ad84605ec0cc005243432