Fix shadow variable warning in ESTreeIRGen-expr.cpp
Summary:
## Instructions about RACER Diffs:
This diff fixes a 'Shadow Variable Warning' issue identified by Quality Insight from [Monetization codehub](https://fburl.com/quality/wmkbc0si):
 - Accept and ship the diff (racer does not do auto land)
 - [**Recommended**] Commandeer and land this diff.

If you are happy with the changes, please accept and ship, or commandeer it if minor edits are needed.
If you are not happy with the changes, please comment on the diff with clear actions and send it back to the author. Racer will pick it up and regenerate.
If you really feel the Racer is not helping with this change (alas, some complex changes are hard for AI) feel free to abandon this diff.

This diff was pre-created by Racer AI agent for your convenience on top of T228185350. More details about the program see this [post](https://fb.workplace.com/groups/1177219810659335/permalink/1198503531864296/). For questions or suggestions please post in [RACER Feedback and Q&A](https://fb.workplace.com/groups/2477474979269093) group.
## Summary:
Fix shadow variable warning in ESTreeIRGen-expr.cpp by renaming the inner 'done' variable to 'throwDone' to avoid shadowing the outer 'done' variable declared at line 1252.
 ---
> Generated by [RACER](https://www.internalfb.com/wiki/RACER_(Risk-Aware_Code_Editing_and_Refactoring)/), powered by [Confucius](https://www.internalfb.com/wiki/Confucius/Analect/Shared_Analects/Confucius_Code_Assist_(CCA)/)
[Session](https://www.internalfb.com/confucius?session_id=0f8811cc-50c9-11f0-872e-e2d3556ca05e&tab=Chat), [Trace](https://www.internalfb.com/confucius?session_id=0f8811cc-50c9-11f0-872e-e2d3556ca05e&tab=Trace)

Reviewed By: GeniusYe

Differential Revision: D76958334

fbshipit-source-id: 950ee853fb851924091dd5b4740f7bdbd4b85f1f