Move compactee evacuation state entirely into EvacAcceptor
Summary:
We currently maintain "evac" pointers in `CompacteeState` that are set
at the end of marking. These pointers allow `EvacAcceptor<true>` to
efficiently check whether an object needs to be evacuated, without
needing to check a separate flag.

However, having this state in the GC is error prone. For example, if the
YG collection runs out of space and blocks on an OG collection, marking
may finish in the middle of the YG collection. That can cause objects
from the compactee to suddenly start getting evacuated halfway through a
YG collection.

To address this, move the pointers into `EvacAcceptor` itself, and make
the decision on whether to do a compaction only at the start of the YG
collection. This keeps control over when exactly to start the compaction
entirely in the YG collection code.

With this, we can re-enable blocking on OG collections when the heap is
exhausted.

Reviewed By: lavenzg

Differential Revision: D76769877

fbshipit-source-id: b85b5a752c09c4ec474f8bec62cd7ef4d234d508